event_be-release/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
mysql-connector-python
passlib
bcrypt
python-dotenv
pydantic[email]
python-jose[cryptography]
pydantic-settings
requests
boto3
aiomysql==0.2.0
httpx==0.28.1
python-jose==3.3.0
python-multipart==0.0.20
python-multipart
razorpay
setuptools>=65.0.0
